@@ -4,7 +4,7 @@
 
     <groupId>org.gautelis</groupId>
     <artifactId>vopn</artifactId>
-    <version>2.5-SNAPSHOT</version>
+    <version>3.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>${project.groupId}:${project.artifactId}</name>
@@ -38,19 +38,19 @@
     </scm>
 
     <properties>
-        <maven.compiler.release>11</maven.compiler.release>
-        <maven.compiler.target>11</maven.compiler.target>
-        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.release>21</maven.compiler.release>
+        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <slf4j.version>2.0.13</slf4j.version>
+        <slf4j.version>2.0.17</slf4j.version>
         <derby.version>10.17.1.0</derby.version>
-        <log4j.version>2.23.1</log4j.version>
+        <log4j.version>2.24.3</log4j.version>
         <log4j-slf4j.version>2.18.0</log4j-slf4j.version>
         <postgresql.version>42.7.5</postgresql.version>
         <antlr4.version>4.13.2</antlr4.version>
-        <axiom.version>1.4.0</axiom.version>
+        <axiom.version>2.0.0</axiom.version>
         <jaxen.version>2.0.0</jaxen.version>
-        <jline.version>3.28.0</jline.version>
+        <jline.version>3.29.0</jline.version>
         <jaxb.api.version>2.3.1</jaxb.api.version>
         <jaxb.runtime.version>4.0.5</jaxb.runtime.version>
         <javax.activation.version>1.1.1</javax.activation.version>
@@ -266,6 +266,21 @@
                         </goals>
                     </execution>
                 </executions>
+            </plugin>
+            <plugin>
+                <artifactId>maven-clean-plugin</artifactId>
+                <version>3.4.0</version>
+                <configuration>
+                    <filesets>
+                        <fileset>
+                            <directory>.</directory>
+                            <includes>
+                                <include>**/*.log</include>
+                            </includes>
+                            <followSymlinks>false</followSymlinks>
+                        </fileset>
+                    </filesets>
+                </configuration>
             </plugin>
         </plugins>
     </build>
